Unit tests e Integration tests

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -32,12 +32,16 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        try {
-            repository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new ResourceNotFoundException(id);
-        } catch (DataIntegrityViolationException e) {
-            throw new DatabaseException(e.getMessage());
+        if (id != null) {
+            try {
+                repository.deleteById(id);
+            } catch (EmptyResultDataAccessException e) {
+                throw new ResourceNotFoundException(id);
+            } catch (DataIntegrityViolationException e) {
+                throw new DatabaseException(e.getMessage());
+            }
+        } else {
+            throw new NullPointerException();
         }
     }
 
@@ -48,10 +52,12 @@ public class UserService {
             return repository.save(entity);
         } catch (EntityNotFoundException e) {
             throw new ResourceNotFoundException(id);
+        } catch (NullPointerException e) {
+            throw new NullPointerException();
         }
     }
 
-    private void updateData(User entity, User user) {
+    private void updateData(User entity, User user) throws NullPointerException{
         entity.setName(user.getName());
         entity.setEmail(user.getEmail());
         entity.setPhone(user.getPhone());
